+++ src/main/java/de/luk/ml/annotation/gui/components/common/ComponentController.java
@@ -36,4 +36,8 @@ public abstract class ComponentController extends BorderPane {
     }
   }
 
+  public void detach(){
+
+  }
+
 }