+++ src/main/java/de/luk/ml/annotation/gui/components/fileselection/SelectOutputFileComponent.java
@@ -12,6 +12,7 @@ import org.apache.commons.io.FilenameUtils;
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
 import java.io.File;
+import java.nio.file.Paths;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 
@@ -33,6 +34,7 @@ public class SelectOutputFileComponent extends ComponentController {
     fc.getExtensionFilters().add(AnnotationFileFilter.getAnyFilter());
     if (!txfFilePath.getText().isEmpty()) {
       fc.setInitialFileName(txfFilePath.getText());
+      fc.setInitialDirectory(Paths.get(txfFilePath.getText()).getParent().toFile());
     } else {
       fc.setInitialDirectory(this.view.root.workingDirectory);
     }
@@ -40,6 +42,13 @@ public class SelectOutputFileComponent extends ComponentController {
     if (outputFile != null) {
       txfFilePath.setText(outputFile.getAbsolutePath());
     }
+  }
+
+  @FXML
+  private void refreshName(){
+    LocalDateTime currentTime = LocalDateTime.now();
+    String filename = currentTime.format(DateTimeFormatter.ofPattern("'entityRecording'-yyyy-MM-dd-HH-mm-ss'.csv'"));
+    this.txfFilePath.setText(FilenameUtils.concat(this.view.root.workingDirectory.getAbsolutePath(), filename));
   }
 
   @PostConstruct
@@ -50,9 +59,7 @@ public class SelectOutputFileComponent extends ComponentController {
   @Override
   public void setView(ViewController view) {
     super.setView(view);
-    LocalDateTime currentTime = LocalDateTime.now();
-    String filename = currentTime.format(DateTimeFormatter.ofPattern("yyyy_MM_dd__kk_mm__'annotations.csv'"));
-    this.txfFilePath.setText(FilenameUtils.concat(this.view.root.workingDirectory.getAbsolutePath(), filename));
+    this.refreshName();
   }
 
   @PreDestroy