+++ src/main/java/de/luk/ml/annotation/gui/views/recording/RecordAnnotationsView.java
@@ -28,4 +28,8 @@ public class RecordAnnotationsView extends ViewController {
     return "Recording";
   }
 
+  @Override
+  public void detach() {
+    this.triggerAnnotationComponent.detach();
+  }
 }