+++ src/main/java/de/luk/ml/annotation/gui/views/review/ReviewAnnotationView.java
@@ -4,6 +4,7 @@ import de.luk.ml.annotation.annotations.AnnotationFile;
 import de.luk.ml.annotation.annotations.AnnotationFileLoader;
 import de.luk.ml.annotation.gui.components.fileselection.SelectAnnotationFile;
 import de.luk.ml.annotation.gui.components.review.AnnotationStats;
+import de.luk.ml.annotation.gui.components.review.CreateEAF;
 import de.luk.ml.annotation.gui.components.review.UpdateStartTime;
 import de.luk.ml.annotation.gui.views.common.ViewController;
 import javafx.event.ActionEvent;
@@ -13,7 +14,7 @@ import javafx.geometry.Pos;
 import javafx.scene.control.Button;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Priority;
-import javafx.scene.layout.TilePane;
+import javafx.scene.layout.VBox;
 
 import javax.annotation.PostConstruct;
 import javax.inject.Inject;
@@ -39,8 +40,11 @@ public class ReviewAnnotationView extends ViewController {
   @Inject
   private AnnotationFileLoader fileLoader;
 
+  @Inject
+  private CreateEAF createEAF;
+
   @FXML
-  private TilePane tilePanel;
+  private VBox vbxPanel;
 
   @PostConstruct
   private void init() {
@@ -69,12 +73,20 @@ public class ReviewAnnotationView extends ViewController {
   @Override
   public void attach() {
     super.attach();
-    if (Objects.nonNull(this.annotationFile)) {
-      this.loadFile(null);
+    if (Objects.nonNull(this.annotationFile) && this.annotationFile.getFile().exists()) {
+      this.selectAnnotationFile.setFilePath(this.annotationFile.getFile().getAbsoluteFile());
+      this.loadAnnotationFile();
+    } else {
+      this.annotationFile = null;
+      this.clearContent();
     }
   }
 
   private void loadFile(ActionEvent action) {
+    this.loadAnnotationFile();
+  }
+
+  private void loadAnnotationFile() {
     if (Objects.nonNull(this.annotationFile)) {
       this.annotationFile.removeChangeListener(this::onChange);
     }
@@ -100,13 +112,19 @@ public class ReviewAnnotationView extends ViewController {
     this.setContent();
   }
 
+  private void clearContent() {
+    this.vbxPanel.getChildren().clear();
+  }
+
   private void setContent() {
-    if (this.tilePanel.getChildren().isEmpty()) {
-      this.tilePanel.getChildren().add(this.stats);
-      this.tilePanel.getChildren().add(this.updateStartTime);
+    if (this.vbxPanel.getChildren().isEmpty()) {
+      this.vbxPanel.getChildren().add(this.stats);
+      this.vbxPanel.getChildren().add(this.updateStartTime);
+      this.vbxPanel.getChildren().add(this.createEAF);
     }
     this.stats.setAnnotations(this.annotationFile.getAnnotations());
     this.updateStartTime.setAnnotationFile(this.annotationFile);
+    this.createEAF.setAnnotationFile((this.annotationFile));
   }
 
   public String getTitle() {