+++ src/main/java/de/luk/ml/annotation/recorder/AnnotationRecorder.java
@@ -46,7 +46,9 @@ public class AnnotationRecorder {
     }
     double started = (currentActiveStart - startTime) / 1000.0;
     double ended = (System.currentTimeMillis() - startTime) / 1000.0;
-    output.write(String.format(Locale.US, "\n%.3f;%.3f;\"" + annotation.getName() + "\"", started, ended));
+
+
+    output.write(String.format(Locale.US, "\n%.3f;%.3f;\"", started, ended).replaceAll(",", ".") + annotation.getName() + "\"");
     output.flush();
     this.currentActiveStart = 0;
     this.currentActive = null;