+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 1.2 (08.03.2017)
+* **Gui:** Set initial directory of the output file selection dialog
+* **Gui:** Change the button text from "Start" to "Next"
+* **Workflow** Use any key to trigger the annotation (except ESC, numbers or Enter)
+* **Bug:** Replace ',' with '.' in decimal numbers
 # 1.1 (09.02.2017)
 * **Gui:** Add version number to status line
 * **Gui:** Set window size for non i3wm users